Fix: ERR | WorkspaceTest | testCreateWorkspace (#1407) 1. Refactored code to create a dedicated WorkspacePage. 2. Updated and executed related test cases.
